+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PixPullSubscription resource
 - PixPullRequest resource
 - IssuingStockRule resource
+- PixFraud.Log resource
 
 ## [0.16.0] - 2026-06-01
 ### Added